% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_FUNCS_1.5.R
\name{add_MonthGroup}
\alias{add_MonthGroup}
\title{Add Month groups}
\usage{
add_MonthGroup(dt, monthgroups_of = 3)
}
\arguments{
\item{dt}{data.table}

\item{monthgroups_of}{(2 or 3, default is 3) if 3 groups are ("Oct - Dec","Jan - Mar","Apr - Jun","Jul - Sep"). if 2, groups are ("Oct-Nov","Dec-Jan","Feb-Mar","Apr-May","Jun-Jul","Aug-Sep").}
}
\value{
dt with new column Month
}
\description{
adds column to data.table for Month group based on Date column
}
\details{
If input data.table has column Date than creates a new column for the Month group as a factor.
}
