% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_FUNCS_1.5.R
\name{add_NAValues}
\alias{add_NAValues}
\title{Add NA's to data serise}
\usage{
add_NAValues(dt, by = "DateTime")
}
\arguments{
\item{dt}{data.table}

\item{by}{time increment to compare to complete series'DateTime' or 'Date'. DateTime assumes 15 minute inteval}
}
\value{
dt with NA's for missing data points
}
\description{
adds NA's for values that are missing from data serise
}
\details{
assumes even intervals of 15 minutes. No duplicates
}
