% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_FUNCS_1.5.R
\name{add_WY}
\alias{add_WY}
\title{Add column for Water Year}
\usage{
add_WY(dt)
}
\arguments{
\item{dt}{data.table}
}
\value{
dt with new column WY
}
\description{
adds column to data.table for Water Year based on Date column
}
\details{
If input data.table has column Date than creates a new column for the Water year (ie 10/01 to 9/30)
}
