% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_FUNCS_1.5.R
\name{add_WYDay}
\alias{add_WYDay}
\title{Add column Day of Water Year}
\usage{
add_WYDay(dt, time = FALSE)
}
\arguments{
\item{dt}{data.table}

\item{time}{Include additional column for wy Day and time fraction}
}
\value{
dt with new column (or two new columns)
}
\description{
Add a column to data.table for day number of water year (1-365) based on Date column
}
\details{
If input data.table has column Date than creates a new column for the Water year day number (ie 10/01(WY-1) is 1, 9/30(WY) is 365/366). Uses yday()
}
