% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDECheck_1.2.R
\name{CDECheck}
\alias{CDECheck}
\title{CDEC Check}
\usage{
CDECheck(
  stations = NULL,
  parameters = "All",
  range = NULL,
  duration = 4,
  graphs = TRUE,
  dt = NULL,
  RedTide = TRUE,
  ErrorCheck = FALSE
)
}
\arguments{
\item{stations}{list of Station codes. Default is all WQES CDEC stations. Other options include station Region ("North Delta", "South Delta", "Central Delta", "Rock Slough")}

\item{parameters}{List of parameters. Default is "All" ("Flow","Temp","SpCond","DO","pH","Turb","Chlor")}

\item{range}{Default is current date time or specify an end date time (or a start and end date time) Use MkDate for specifying date time.}

\item{duration}{Specify number of days prior to end date to download data for. (Ignored if range includes two values, start and end dates) Problems may occure with start or end times at midnight 00:00}

\item{graphs}{TRUE generates a PDF plot containing all station and parameter combos. (can take several seconds to finish)}

\item{dt}{To skip the data download step, supply a data.table. If stations and parameters are unspecified, will graph all station/parameter combos in dt}

\item{ErrorCheck}{Use TRUE to print during error checking.}
}
\value{
data.table of downloaded CDEC data. When graphs = TRUE also creates a PDF of plotted data.
}
\description{
Download and plot all WQES data from CDEC. Saves plot to Water Quality Evaluation Section/Tools/Rstuff/CDECheck/ (or working driectory if not connected to //pcdistfs1/share)
}
\details{
Download and plot data for daily CDEC checks. data table is returned from function and PDF plot is saved to "Water Quality Evaluation Section/Tools/Rstuff/CDECheck/"
         StationOrder=c("SOI","SXS","MIR","BLP","FAL","FCT","TSL","MOK","OSJ","BET","HOL","ORQ","RSCC","DRB","WDC","HCHM","OBI","HLT","TRN","VCU","MRX","MHO","MRU","WCI","ORI","GLC","ODM","ORM","TWA","GLE","SGA","TPI","PDC","ORX","OH1","SJD")
}
