% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateCheck_1.1.R
\name{DateCheck}
\alias{DateCheck}
\title{Looking for missing times and duplicate time stamps}
\usage{
DateCheck(Xdt, Fix = FALSE, PDT = FALSE, Diagnos = FALSE)
}
\arguments{
\item{Xdt}{data.table with column "DateTime"}

\item{Fix}{should sequence be replaced and NULLS inserted?}

\item{PDT}{for input of daylight savings}

\item{Diagnos}{I know know}
}
\value{
original data.table (if FIX, missing times are inserted with NULL for values)
}
\description{
old function, needs to be revisited
}
\details{
function needs work
}
