% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MkDate_1.3.R
\name{DateFormat}
\alias{DateFormat}
\title{Determine DateFormat}
\usage{
DateFormat(DateForm)
}
\arguments{
\item{DateForm}{Singe Date Time value as character}
}
\value{
Returns a Date Format to be used by POSIXct format parameter.
}
\description{
Sets expected Date/Time format from list of pre set common types
}
\details{
Somethng for later
}
