% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dmelt_Dcast.R
\name{Dcast}
\alias{Dcast}
\title{Cast standard WQES data.table}
\usage{
Dcast(dt, value.var = "value")
}
\arguments{
\item{dt}{input data.table}

\item{value.var}{(default="value") column name to cast (ie populate columns of Station.Parameter)}
}
\value{
a wide table format data.table
}
\description{
Convert Long table to Wide table. Converts from standardized column naming of (Date/DateTime), Station, Parameter, value to columns named as [station].[parameter]
}
\details{
use this function to convert a data.table from GetData or Dmelt into a wide format.
}
