% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dmelt_Dcast.R
\name{Dmelt}
\alias{Dmelt}
\title{Melt standard WQES data.table}
\usage{
Dmelt(dt, DateTime_col = NULL, type = FALSE, importer = FALSE, CDEC = FALSE)
}
\arguments{
\item{dt}{input data.table}

\item{DateTime_col}{to specify either DateTime or Date, if blank assumes first column name is Time column to melt by}

\item{type}{if TRUE adds Type column based on input column names (ie PDC.Temp.Min)}

\item{importer}{if input file is from WQES_Importer use TRUE to remove duplicated rows the occure due to multi-lined comments. (comments will be removed)}
}
\value{
a long format data.table with columns (Date/DateTime), Station, Parameter, variable, value
}
\description{
Convert Wide table to long table. Converts from standardized [station].[parameter] column naming to Station, Parameter, value column names
}
\details{
use this function to convert a data.table back into long format after being cast into wide format from Dcast, Dread (or other function within Dtoolbag)
}
