% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dmerge.R
\name{Dmerge}
\alias{Dmerge}
\title{Combine data.tables}
\usage{
Dmerge(ldt, dt2 = NULL)
}
\arguments{
\item{ldt}{First data.table (or list of data.tables, in which case all tables will be combined)}

\item{dt2}{Second data.table. Should not be a list.}
}
\value{
Single data.table. All columns are considered Keys, so rows with all matching columns will be combined.
}
\description{
Combine data.tables that have come from GetData
}
\details{
Merges two data tables on shared columns
}
