% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dplotsave_1.1.R
\name{Dplotsave}
\alias{Dplotsave}
\title{save R plots as .png}
\usage{
Dplotsave(plotname, plot, size = "medium", folder = "plots", PlotFile = NULL)
}
\arguments{
\item{plotname}{file name (include extra folders to embed within working directory)}

\item{plot}{R plot. (not tested with non-ggplot plots)}

\item{size}{(Default is 'medium') (c(Width,Height) in inches) or use options: \cr
medium - (7.5x4) fits two plots per page \cr
large - (10x6.5) fits one plot per page landscape \cr
tall - (7.5x7.5) fits one plot per page with space for paragraph \cr
taller - (7.5x9.5) fits one plot per page portrate with space for caption}

\item{folder}{(default = 'Plots')folder Specify a folder within working directory}
}
\value{
absolute path to saved plot file
}
\description{
Save R plots to working directory (or ...?) using set sizes that match well in word docs
}
\details{
Saves a plot into your working directory using preset sizes for uniformity
}
