% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dread_1.5.R
\name{Dread}
\alias{Dread}
\title{Read in common files for WQES}
\usage{
Dread(
  file,
  time15min = TRUE,
  quality = TRUE,
  comments = TRUE,
  station = NULL,
  from = NULL,
  nRstuff = FALSE,
  filter = NULL,
  ErrorCheck = FALSE
)
}
\arguments{
\item{file}{file name and path you want to read.}

\item{time15min}{TRUE replaces times to an even 15 minute interval (ie 10:16:18 -> 10:15:00)}

\item{quality}{(TRUE/FALSE) Should data quality column be kept? Default is TRUE}

\item{comments}{(TRUE/FALSE) Should comments column be kept? Default is TRUE}

\item{station}{Set station code for files not in the WQES referances.}

\item{from}{(one of: "file","HYD","WDL",or "Importer") Automagically determined in function.}

\item{nRstuff}{If TRUE, file path starts ar Rstuff/data_files/. IF FALSE file name should be absolute}
}
\value{
data.table with column names in the form DateTime,(StationCode).(ParameterAbreviation)
}
\description{
Reads in common file types and converts them to a standard format data.table (column names: DateTime, StationCode.Parametercode)
}
\details{
If crashes try using fread()
}
