% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DreadDiscrete_1.1.R
\name{DreadDiscrete}
\alias{DreadDiscrete}
\title{Read Discrete lab data}
\usage{
DreadDiscrete(file, simple = TRUE, filter = NULL, ErrorCheck = FALSE)
}
\arguments{
\item{file}{file name and path you want to read.}

\item{simple}{Default is TRUE, simplifies table}

\item{filter}{Default is NULL, (or all) list of parameters}
}
\value{
data.table with column names in the form DateTime,(StationCode).(ParameterAbreviation)
}
\description{
Reads Discrete lab data
}
\details{
If crashes try using fread()
}
