% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DreadFolder_1.1.R
\name{DreadFolder}
\alias{DreadFolder}
\title{Read all files in a folder}
\usage{
DreadFolder(folder = "Combine", nRstuff = TRUE, ErrorCheck = FALSE)
}
\arguments{
\item{folder}{folder name to read. (default is 'Combine' in Rstuff/data_files).}

\item{nRstuff}{is folder relative to Rstuff/data_files/? (use TRUE) Else, relative to working directory}
}
\value{
dt with all data
}
\description{
Created so files could be downloaded from new WDL, placed in a folder, and combined in R. Dreadfolder() will read and combine all files in user folder Documents/Rstuff/data_files/Combine
}
\details{
Files must work with function Dread, which generally means file was downloaded from WDL or Hydstra.
 If file contains multiple stations or parameters, Quality codes will be erased (v1.1).
 Note: Dread default of time15min=TRUE is used.
}
