% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DreadKorCal.R
\name{DreadKorCal}
\alias{DreadKorCal}
\title{Read KOR calibration data files}
\usage{
DreadKorCal(file, parameters = NULL, range = NULL)
}
\arguments{
\item{file}{file name and path you want to read.}

\item{parameters}{parameters to filter}

\item{range}{date range to filter}
}
\value{
list of two data.tables, Kor Cals and Cal points
}
\description{
Reads Kor calibration data files
}
\details{
I dont know
}
