% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRoll_1.1.R
\name{DRoll}
\alias{DRoll}
\title{Rolling window function}
\usage{
DRoll(
  dt,
  days = 30,
  funs = c("mean", "max", "min", "sum", "sd"),
  align = "center",
  merge = FALSE
)
}
\arguments{
\item{dt}{Input data.table}

\item{days}{number of days to apply rolling window (default is 30)}

\item{funs}{functions to apply rolling window to. default: c("mean","max","min","sum","sd")}

\item{align}{for froll (default 'center')}

\item{merge}{use TRUE to return orignal table with added columns for rolling funcs}

\item{ErrorCheck}{Use TRUE to print during error checking.}
}
\value{
data.table of Common stats for rolling funs by input groups
}
\description{
Apply a rolling function by groups for data.table
}
\details{
returns a rolling window with funs applied data.table for a rolling window of number of days. First calculates funs for Daily values and then aggrgates by rolling window
}
