% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtSummary.R
\name{dtSummary}
\alias{dtSummary}
\title{Summarize a data.table as made by Dmelt(Dread)}
\usage{
dtSummary(dt)
}
\arguments{
\item{dt}{a data.table}
}
\value{
Returns data.table of summary
}
\description{
Summarize a data.table as made by Dmelt(Dread) to show Max's Min's Range and value count
}
\details{
Files should be melted files ran through Dread
}
