% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DupCheck.R
\name{DupCheck}
\alias{DupCheck}
\title{Find duplicates}
\usage{
DupCheck(dt)
}
\arguments{
\item{dt}{input data.table, From GetData()}
}
\value{
dtdups: a data.table list conatining $dups $points $mean
}
\description{
Find duplicates in data.tables for DateTime, Station, Parameter~ value
}
\details{
For tables that include DateTime, Station, Parameter, and value.
         Returns list dtdups:
         $dups: Combinations of DateTime, Station, and Parameter that produced a duplicate
         $points: all points that are duplicates of DateTime, Station, and Parameter
         $mean: Fixed dt with duplicates removed by taking average of duplicate points. (NA's removed)
}
