% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Files4R_1.1.R
\name{Files4R}
\alias{Files4R}
\title{Find file names}
\usage{
Files4R(path = "", string = "*", nRstuff = FALSE)
}
\arguments{
\item{path}{Project name (or sub directory with either working directory or Rstuff/data_files)}

\item{string}{regular expression used to search file names. Default is "*" which returns all files. (use "*.csv" to return all .csv files)}

\item{nRstuff}{(Default is FALSE) search path srating in working directory. if TRUE: search path starting in Rstuff/data_files.}
}
\value{
vector of file names matching string within folder specified by path
}
\description{
Find file names in either working directory or Documents/Rstuff/data_files
}
\details{
Used to search and get file names and absolute paths.
}
