% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydstra_funcs_1.1.R
\name{get_HYCSV}
\alias{get_HYCSV}
\title{Read recent output from HYCSV in Hydstra drive}
\usage{
get_HYCSV(num = 1)
}
\arguments{
\item{num}{Default is 1. Number of files to read, IDK, i'm making this up as I go}
}
\value{
Returns latest num files in Q:Users/DCOLVIN/TEMP/TEMP
}
\description{
Read and import recent export from HYCSV
}
\details{
After running HYCSV, grabs the latest file. then Use Dread
}
