% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData_1.5.R
\name{GetData}
\alias{GetData}
\title{Download Water Quality Data}
\usage{
GetData(
  stations,
  parameters,
  range = NULL,
  source = "WDL",
  duration = NULL,
  stationAS = "StationCode",
  POR = FALSE,
  replace = TRUE,
  type = "15Minute",
  time15min = TRUE,
  ErrorCheck = FALSE,
  report = FALSE,
  dtGet = NULL
)
}
\arguments{
\item{stations}{a vector of Station Codes that match StationCode in tbl_Stations (or see dtGet)}

\item{parameters}{a character vector of parameters that match ParamR in tbl_Parameters (Temp, DO, DOsat, pH, SpCond, Turb, Chlor, Flow, Stage(only avaiable from CDEC)) Default is "All".
For source="WDL", field HydstraNum in tbl_Stations must be present for specified ParamR row. For source="CDEC", CDECNum in tbl_Parameters must be present, and for source="USGS", USGSNum must be present.}

\item{range}{date range. For best results specify Start and End date with time and use MkDate() (default=NULL and if WDL will return current year, CDEC returns past 10 days, USGS returns past 365 days).}

\item{source}{Source to download data ("CDEC", "WDL", or "USGS") Default is "CDEC". WDL requires field ID_HydstraWQ in tbl_Stations for given StationCode (or ID_HydstraFlow for Flow data). USGS requires ID_USGS for given StationCode. (or see dtGet)}

\item{duration}{used to specify CDEC data range when range is not entered}

\item{POR}{(deafult FALSE) Download Period Of Record from WDL as daily average values. Note: any day with one or more 170 or 151 points will not be included}

\item{replace}{(default TRUE) For WDL and USGS, new download files will replace prious ones. If False, will first check for existing file and not replace it.}

\item{type}{(Default "15Minute" or "Raw" for WDLnew) For USGS other type is "Daily" but is untested.. For New WDL 15minute is Raw or use Daily_Mean.}

\item{time15min}{(unused)}

\item{ErrorCheck}{(Default is FALSE)}

\item{report}{(Default FALSE) Shows table of downloaded files and failed files at end.(dtGet)}

\item{dtGet}{to directly specify the Station, Parameter, and ID combinations and bypass setting stations,parameters.
if source = WDL, dtGet must have columns (StationCode, ParamR, ID_Hydstra, Year)
if source = CDEC, (dtGet cannot be entered)
if source = USGS, dtGet must have columns (StationCode, ParamR, ID_USGS, Start, END)}
}
\value{
Returns a data.table long format of downloaded data.
}
\description{
Downloads water quality data for stations and parameters that have been set up in WQES_database tbl_Stations and tbl_Parameters. Sources include CDEC, WDL, and USGS
}
\details{
For WDL and USGS, raw files are downloaded to Documents/Rstuff/data_files. Post processing of data is required, use DupCheck() to find duplicates.
}
