% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Files4R_1.1.R
\name{getMostRecentFile}
\alias{getMostRecentFile}
\title{Return most recent file}
\usage{
getMostRecentFile(folderPath)
}
\arguments{
\item{folderPath}{the path to a folder}
}
\description{
Return most recent file in a folder from specified path.
}
