% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPath2Rstuff.R
\name{GetPath2Rstuff}
\alias{GetPath2Rstuff}
\title{Set Path2Rstuff}
\usage{
GetPath2Rstuff()
}
\value{
The path way to Rstuff. If path cannot be determined defaults to getwd()
}
\description{
Use by Path2Rstuff<-GetPath2Rstuff()
Assumes computer has path Documents/(User Name)/Rstuff
Allows for reading and writing data files to Rstuff/data_files/(project names)
}
