% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_FUNCS_1.5.R
\name{GodinFilter}
\alias{GodinFilter}
\title{Add tidal filtered value}
\usage{
GodinFilter(
  dt,
  addNA = TRUE,
  version = "Huston",
  partial = FALSE,
  fill = NA,
  na.rm = FALSE
)
}
\arguments{
\item{dt}{data.table}

\item{version}{Huston or Swift (oringally was Swift version, now is Huston version)}

\item{NAlimit}{integer numbr of allowable NA's in averageing window}
}
\value{
dt with new column tfvalue
}
\description{
adds column to data.table for tidally filtered values column based on the GodinFilter
}
\details{
Smoothing function applied for each Station and Parameter in data.table. Assumes input data has no missing 15 minute intervals. Adds column tfvalue as the average of (center weighted 24 hour rolling average + previous 25 hour average + future 25 hour average).  Problems occure with NA or NULL values, or missing 15 minutes intervals. Use fill="extend" to avoid gaps. Tidal filter will also elliminate short duration pulse signals.
}
