% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MkDate_1.3.R
\name{MkDate}
\alias{MkDate}
\title{Convert common timstamps to POSIXct}
\usage{
MkDate(
  DateTime,
  asDate = FALSE,
  time15min = FALSE,
  sequence = FALSE,
  BY = "15 min",
  OFFset = "",
  PDT = FALSE,
  DateForm = "",
  ErrorCheck = FALSE,
  UTCcheck = TRUE
)
}
\arguments{
\item{DateTime}{vector of dates, need to be an a format already in DateFormat
If typing numbers, must be characters (ie use "202010301015" for 10/30/2020 10:15)}

\item{asDate}{Converts to Date only (no time)}

\item{time15min}{Make times even intervals of 15 minutes  (:00 :15 :30 :45)}

\item{sequence}{Makes a sequence using BY}

\item{BY}{Time interval to be used by sequence}

\item{OFFset}{I don't know}

\item{PDT}{Use TimeZone "America/Los_Angeles" (aka daylight savings)}

\item{DateForm}{Specify the date formate (if blank DateForm is set by function DateFormat)}
}
\value{
Converted POSIXct timestamps
}
\description{
convert common date and time formats to POSIXct, timezone "ETC/GMT+8"   (aka pst)
}
\details{
If crashes try using fread(). (sometimes fread will assume values are in UTC, and cause a shift when using MkDate...)
}
