% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydstra_funcs_1.1.R
\name{set_hycsv}
\alias{set_hycsv}
\title{Replace HYCSV history file with new export criteria}
\usage{
set_hycsv(
  stations,
  parameters,
  range,
  source = "A",
  type = "POINT",
  quality = "No",
  interval = "15"
)
}
\arguments{
\item{stations}{a vector of Station Codes that match StationCode in tbl_Stations}

\item{parameters}{a character vector of parameters that match ParamR in tbl_Parameters with corresponding HydstraCode (Temp, DO, DOsat, pH, SpCond, Turb, Chlor, Flow, Stage).}

\item{range}{date range.(use MkDate)}

\item{source}{hydstra file source. Default is "A" for archive}

\item{type}{hydstra export type. Default is "POINT"}

\item{quality}{default is "No" (or use"Yes")}

\item{interval}{default is 15. Or use Daily.}
}
\value{
Returns dtGet, data.table of stations and parameters sent to HYCSV
}
\description{
By replacing HYCSV history, you can open HYCSV in hydstra, select "Use previous export" and the new export criteria will populate the user form.
}
\details{
Function re-writes the csv file located in "Q:/Users/DCOLVIN/TEMP/PRM/HYCSV.PRM" with new hydstra station codes, parameter codes and date range for quick population of HYCSV.
Each combo of station and parameter is set into HYCSV
Can only do 6 times series at a time. (ie stations * parameters <= 10)
}
