% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydstraWebServices_1.1.R
\name{WDL_SiteTraceData}
\alias{WDL_SiteTraceData}
\title{Download data from WDL using web services}
\usage{
WDL_SiteTraceData(
  hydstra_id,
  param = "",
  range = NULL,
  json = FALSE,
  Trace = NULL,
  type = "RAW"
)
}
\arguments{
\item{param}{hydstra web service param name as character (or enter a Trace)}

\item{range}{two variable vector for start and end date. Enter as character with format YYYYMMDDHHMM (can be entered as returned from MkDate)}

\item{json}{if true returns json file. (or, is a list that can be parsed as a json?)}

\item{Trace}{hydstra trace as character. Overirdes param if entered. (a trace in hydstra webservices is the param and duration. eg WaterTemp.Raw are point values)}

\item{type}{hydstra duration or processing. Default is "RAW" which is instantanious point data. Or enter MEAN for daily means ending at midnight}

\item{hystra_id}{should be a valid hydstra id (eg B9539000)}
}
\value{
either a table/list of timeseries data (trace), or a json of trace
}
\description{
Download time series data (Traces) for individual site over specied range from WDL.
}
\details{
Download data from WDL hydstra web services. Data is specified from hydstra_id, param or Trace, and range.
         If a Trace is entered, that trace will be downloaded. otherwise trace is specified as param + type. Default type is RAW, or point data as stored in Hydstra
}
