% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydstraWebServices_1.1.R
\name{WDL_SiteTraces}
\alias{WDL_SiteTraces}
\title{See available traces for individual WDL/Hydstra site.}
\usage{
WDL_SiteTraces(hydstra_id, json = FALSE)
}
\arguments{
\item{json}{if true returns json file. (or, is a list that can be parsed as a json?)}

\item{hystra_id}{should be a valid hydstra id (eg B9539000)}
}
\value{
either a table/list of avialable traces, or a json of aviable traces
}
\description{
See available time series data (Traces) for individual sites to download from WDL.
}
\details{
use this function to see what traces (time series data) are aviable for a site through WDL
}
