% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WQES_tbls.R
\name{WQES_tbls}
\alias{WQES_tbls}
\title{Read WQES tables}
\usage{
WQES_tbls(update = FALSE)
}
\arguments{
\item{update}{if TRUE, copies files tbl_Parameters.txt and tbl_Stations.txt from Share drive folder Z://pcdistfs1/share/Water Quality Evaluation Section/Tools/Rstuff/data_files/db_tables to user's current R library Dtoolbag folder. If Error occures, verify connection to //pcdistfs1/share/. Alternatively, copy table files from Water Quality Evaluation Section/Tools/Rstuff/data_files/db_tables and paste into your R library Dtoolbag folder (replacing existing files)}
}
\value{
list of tbl_Stations and tbl_Parameters from Package
}
\description{
reads tbl_Stations and tbl_Parameters (first from Z drive, if available, or from R library)
}
\details{
Loads the tbl_Stations and tbl_Parameters from the R library Dtoolbag folder and returns as a list with $Stations and $Parameters.
}
